pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansible-core>=2.20.1",
    "fastapi[standard]>=0.128.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
